+++ GameGMD/Enemies/EnemyData.cs
@@ -19,7 +19,7 @@ namespace GameGMD
             physics.MoveDownBy(distance);
         }
 
-        public void UpdatePosition(double X, double Y)
+        public void UpdatePosition(float X, float Y)
         {
             physics.UpdatePosition(X, Y);
         }
