+++ GameGMD/Enemies/Pinkie.cs
@@ -26,9 +26,9 @@ namespace GameGMD.Flyweight
             this.distance = distance;
         }
 
-        public void UpdatePosition(double X, double Y)
+        public void UpdatePosition(float X, float Y)
         {
-            Enemy.UpdatePosition(X+0.01, Y);
+            Enemy.UpdatePosition(X+0.01f, Y);
         }
 
         public bool ShouldDestroy()
